Findal draft of Stats section Color variables postponed till the responsive overhaul
