+++ project/src/global.css
@@ -4,10 +4,10 @@
 
 :root {
     /* Body Vars */
-    --gradient-bg-color-1: 231deg 100% 10%;
-    --gradient-bg-color-2: 256deg 62% 10%;
+    --gradient-bg-color-1: 231deg 100% 6%;
+    --gradient-bg-color-2: 256deg 62% 12%;
     --gradient-bg-color-3: 243deg 62% 9%;
-    --gradient-bg-color-4: 240deg 62% 10%;
+    --gradient-bg-color-4: 240deg 62% 3%;
 
     /* Text Vars */
     --text-primary: 0deg 0% 99%;
